+++ Alloy/commands/compile/parsers/Ti.UI.Vertical.js
@@ -1,0 +1,2 @@
+// This is an alias for Ti.UI.VerticalLayout
+module.exports = require('./Ti.UI.VerticalLayout');